requirements.txt
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
fake-useragent>=1.1.0
jieba>=0.42.1
streamlit>=1.20.0
playwright>=1.40.0
